Lightning is here
I implemented Lightning. The spell is instantiated as an ASpell and
begins to search for the best AActor around. It uses a cone to do so.
If it found an actor restart the search for this actor but use a box
instead. up to 4 times.
If search is done, set up a timer and start shooting lightning and apply
damage each tick to the actors that were previously found (programmer
defined amount.)